+++ src/helper/configs/aruntime/project/sylius.go
@@ -1,0 +1,13 @@
+package project
+
+func init() {
+	RegisterDockerConfGenerator("sylius", MakeConfSylius)
+}
+
+func MakeConfSylius(projectName string) {
+	MakePhpDockerfile(projectName)
+	MakeDBDockerfile(projectName)
+	MakeRedisDockerfile(projectName)
+	MakeScriptsConf(projectName)
+	MakeClaudeDockerfile(projectName)
+}